renamed plugins to custom prepare for merge
